move elected validators info fetch to separate factory
