+++ fearless/Modules/Staking/Model/PreparedNomination.swift
@@ -1,0 +1,8 @@
+import Foundation
+
+struct PreparedNomination {
+    let amount: Decimal
+    let rewardDestination: RewardDestination
+    let fee: Decimal
+    let targets: [SelectedValidatorInfo]
+}
