+++ fearless/Modules/Staking/Model/SelectedValidatorInfo.swift
@@ -1,0 +1,6 @@
+import Foundation
+
+struct SelectedValidatorInfo {
+    let address: String
+    let identity: AccountIdentity?
+}
